sparkler: use fb.h put_pixel helpers
discards draw_pixel(), introduces helpers.h and a convenience
function for bounds checking and oob extermination.  Move makergb
to helpers.h, draw.h gets removed in a later commit.
