+++ src/modules/sparkler/draw.h
@@ -5,15 +5,6 @@
 
 #include "fb.h"
 
-/* helper for scaling rgb colors and packing them into an pixel */
-static inline uint32_t makergb(uint32_t r, uint32_t g, uint32_t b, float intensity)
-{
-	r = (((float)intensity) * r);
-	g = (((float)intensity) * g);
-	b = (((float)intensity) * b);
-
-	return (((r & 0xff) << 16) | ((g & 0xff) << 8) | (b & 0xff));
-}
 
 static inline void draw_pixel(fb_fragment_t *f, int x, int y, uint32_t pixel)
 {
